Relocate recursion detection to visitor.visit().
This commit removes the  recursion detection logic from
`visitor.visitMap()`, `visitor.visitPtr()`, and `visitor.visitSlice()`
methods and places it on the top of `visitor.visit()` method to have a
single point of recursion detection and recursion marker printing.

Adding `Value.CanNil()` method is required to detect whether it is safe
to call `Value.IsNil()` on all possible kinds of values without causing
the code to panic.

Also an if branch was placed inside the `visitor.enter()` method to exit
the method if the value is an empty interface. That was required to keep
the interface rendering tests from failing and to keep the behaviour
backward-compatible.