Let only concurrency detection in visitor.enter().
The purpose of this commit is to refactor `visitor.enter()` method to
narrow down its purpose to concurrency detection only. Previous nil
value rendering was relocated to the types that specifically require nil
rendering (ptr, slice, and map). Nil interface rendering was left intact
as it requires some specifics when rendering ambiguous static types.

As the remaining logic in `visitor.enter` now requires to receive a
pointer from the given type, `Value.canNil()` method was replaced with
`Value.canPointer()` method to check if it is safe to acquire a pointer
from a value.