Collapse zero-value structs.
This commit changes the printer behavior to render zero-value struct types more
concisely when all of its fields are zero-values.

Empty structs, that is, structs with no fields are not affected by this change,
as doing so would actually increase the output size and lose information.

The zero value "marker" is configurable, much the same as the recusion marker.